feat(backend): add full function of HM0202，HM0402; update Initialize class; update DB-SPEC-V1.md

+++ backend/src/main/java/com/frogsoft/frogsoftcms/controller/v1/request/home/AnnouncementsSetRequest.java
@@ -8,5 +8,5 @@ import lombok.Data;
 @Data
 public class AnnouncementsSetRequest implements Serializable {
   private static final long serialVersionUID = 32974805623333756L;
-  private List<Integer> articleIds;
+  private List<Long> articleIds;
 }
